Update to 3.4 and allow multi-file processing
Changed file from ipython notebook to a .py and updated string function
to 3.4's version.
The program now processes any .txt files found in the current directory
and saves the output as separate files with a suffix.
Declared one function, but it could use at least one more to tighten up
the code